# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HackYeah 2026 - ROPS Social Innovations Vector Search API",
    "description": "Production-ready Chroma DB vector search REST API powered by Google Gemini Embeddings 2 (`gemini-embedding-2`).\n\n### Indexed Dataset: `rops_innovations`\nContains **115 indexed social innovations** from Regional Centre for Social Policy (ROPS Kraków). Use `POST /api/query` to execute semantic vector similarity searches over innovative products, therapies, and solutions for seniors, individuals with disabilities, and social care organizations.",
    "version": "1.0.0",
    "contact": {
      "name": "HackYeah 2026 Engineering Team"
    }
  },
  "servers": [
    {
      "url": "http://ohzlsc0rzubuinsvjuapujke.57.131.158.238.sslip.io",
      "description": "Coolify Production Server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    },
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
    "/api/query": {
      "post": {
        "tags": [
          "Query"
        ],
        "summary": "Semantic search across ROPS innovations",
        "description": "Performs vector similarity search using `gemini-embedding-2` in 3072 dimensions.\n\n**Key Features:**\n- Target collection defaults to `rops_innovations`.\n- Supports natural Polish queries (e.g. *stymulacja pamięci seniora*, *inteligentne czujniki do leków*).\n- Supports Chroma metadata filtering via `where` (e.g. `{\"category_slug\": \"dla-seniorow\"}`).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              },
              "examples": {
                "sensors_and_devices": {
                  "summary": "Search smart devices & sensors for seniors",
                  "value": {
                    "api_key": "YOUR_API_KEY",
                    "collection_name": "rops_innovations",
                    "query": "inteligentne opaski i czujniki dla seniorów",
                    "n_results": 3
                  }
                },
                "sensory_boards": {
                  "summary": "Search sensory boards & dementia therapies",
                  "value": {
                    "api_key": "YOUR_API_KEY",
                    "collection_name": "rops_innovations",
                    "query": "drewniana tablica manipulacyjna dla seniora z demencją",
                    "n_results": 2
                  }
                },
                "category_filtered": {
                  "summary": "Search with category metadata filter",
                  "value": {
                    "api_key": "YOUR_API_KEY",
                    "collection_name": "rops_innovations",
                    "query": "pomoc w rehabilitacji ruchowej",
                    "n_results": 2,
                    "where": {
                      "category_slug": "dla-osob-o-ograniczonej-mobilnosci"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ranked semantic search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                },
                "example": {
                  "status": "success",
                  "collection": "rops_innovations",
                  "results": [
                    {
                      "query": "inteligentne opaski i czujniki dla seniorów",
                      "matches": [
                        {
                          "id": "dla-seniorow__therapy-set",
                          "distance": 0.4188,
                          "document": "Tytuł innowacji: Therapy Set\n\nKategoria: Innowacje dla seniorów\n\nProblem: Ograniczona dostępność terapii zajęciowej dla seniorów...",
                          "metadata": {
                            "title": "Therapy Set",
                            "category_id": 1,
                            "category_name": "Innowacje dla seniorów",
                            "category_slug": "dla-seniorow",
                            "authors": "Joanna Krokosz",
                            "source_url": "https://rops.krakow.pl/innowacje-spoleczne/biblioteka-innowacji-spolecznych/dla-seniorow,therapy-set",
                            "file_zip": "https://rops.krakow.pl/pliki/IS/bibloteka/therapyset.zip",
                            "youtube_video": "",
                            "details_pdf": "[\"https://rops.krakow.pl/mpliki/IS/BIBLIOTEKA_INNOWACJI_SPOECZNYCH/Zasady_wykorzystania_innowacji_MIIS.pdf\"]",
                            "data_source": "rops.krakow.pl",
                            "scrape_timestamp": "2026-10-03",
                            "index": 16
                          }
                        },
                        {
                          "id": "dla-zdrowia-i-medycyny__inteligentny-organizer-do-lekow",
                          "distance": 0.4375,
                          "document": "Tytuł innowacji: Inteligentny organizer do leków\n\nKategoria: Innowacje dla zdrowia i medycyny\n\nRozwiązanie: Narzędzie wyposażone w czujniki...",
                          "metadata": {
                            "title": "Inteligentny organizer do leków",
                            "category_id": 9,
                            "category_name": "Innowacje dla zdrowia i medycyny",
                            "category_slug": "dla-zdrowia-i-medycyny",
                            "authors": "Krzysztof Lingo Wojciech Sójka",
                            "source_url": "https://rops.krakow.pl/innowacje-spoleczne/biblioteka-innowacji-spolecznych/dla-zdrowia-i-medycyny,inteligentny-organizer-do-lekow",
                            "file_zip": "https://rops.krakow.pl/pliki/IS/bibloteka/inteligentnyorganizer.zip",
                            "details_pdf": "[]",
                            "index": 111
                          }
                        }
                      ]
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized (invalid or missing api_key)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error in request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/collections": {
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "List collections",
        "description": "Lists all Chroma DB collections with record counts and indexing metrics.",
        "parameters": [
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "API Authorization Key"
          }
        ],
        "responses": {
          "200": {
            "description": "List of collections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionsResponse"
                },
                "example": {
                  "status": "success",
                  "collections": [
                    {
                      "name": "rops_innovations",
                      "count": 115,
                      "metadata": {
                        "hnsw:space": "cosine"
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Collections"
        ],
        "summary": "Create collection",
        "description": "Explicitly creates a collection. Requires administrative authorization (ADMIN_API_KEY).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCollectionRequest"
              },
              "example": {
                "admin_api_key": "YOUR_ADMIN_API_KEY",
                "name": "rops_innovations",
                "metadata": {
                  "hnsw:space": "cosine",
                  "topic": "rops_social_innovations"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Collection created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleCollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Collection already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Standard API key does not have permission to create collections",
            "content": {
              "application/json": {
                "schema": {
                  "": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminApiKeyAuth": []
          }
        ]
      }
    },
    "/api/collections/{collection_name}": {
      "get": {
        "tags": [
          "Collections"
        ],
        "summary": "Get collection info",
        "description": "Retrieves count and metadata for a specific collection.",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "default": "rops_innovations"
            },
            "description": "Name of the collection"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "API Key"
          }
        ],
        "responses": {
          "200": {
            "description": "Collection details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleCollectionResponse"
                },
                "example": {
                  "status": "success",
                  "collection": {
                    "name": "rops_innovations",
                    "count": 115,
                    "metadata": {
                      "hnsw:space": "cosine"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Collection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Collections"
        ],
        "summary": "Delete collection",
        "parameters": [
          {
            "name": "collection_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Name of collection to delete"
          },
          {
            "name": "api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "API Key"
          },
          {
            "name": "admin_api_key",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Administrative API Key"
          }
        ],
        "responses": {
          "200": {
            "description": "Collection deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Standard API key cannot delete collections",
            "content": {
              "application/json": {
                "schema": {
                  "": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "AdminApiKeyAuth": []
          }
        ],
        "description": "Deletes a collection and all of its records. Requires administrative authorization (ADMIN_API_KEY)."
      }
    },
...
    "/api/documents/upsert": {
      "post": {
        "tags": [
          "Documents"
        ],
        "summary": "Upsert documents (insert or update)",
        "description": "Upserts documents into `rops_innovations`. Automatically generates 3072-dim embeddings using `gemini-embedding-2`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertDocumentsRequest"
              },
              "example": {
                "api_key": "YOUR_API_KEY",
                "collection_name": "rops_innovations",
                "ids": [
                  "dla-seniorow__bawita"
                ],
                "documents": [
                  "Tytuł innowacji: BaWita\n\nKategoria: Innowacje dla seniorów\n\nProblem: Zespoły otępienne..."
                ],
                "metadatas": [
                  {
                    "title": "BaWita",
                    "category_slug": "dla-seniorow",
                    "authors": "Maria Lorenc"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Documents upserted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentActionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Authorization key. Can be supplied in header `X-API-Key`, `Authorization: Bearer <key>`, or inside JSON body as `\"api_key\": \"...\"`"
      },
      "AdminApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-API-Key",
        "description": "Administrative key required for creating and deleting collections. Can also be supplied in JSON body as admin_api_key or query param ?admin_api_key="
      }
    },
    "schemas": {
...
      "QueryResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "success"
          },
          "collection": {
            "type": "string",
            "example": "rops_innovations"
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueryResultItem"
            }
          }
        },
        "required": [
          "status",
          "collection",
          "results"
        ]
      },
      "QueryRequest": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string",
            "example": "inteligentne opaski i czujniki dla seniorów",
            "description": "Search query text in natural language"
          },
          "query_texts": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional batch of search queries"
          },
          "collection_name": {
            "type": "string",
            "default": "rops_innovations",
            "example": "rops_innovations",
            "description": "Target collection (defaults to rops_innovations)"
          },
          "n_results": {
            "type": "integer",
            "default": 5,
            "example": 3,
            "description": "Number of top matching documents to return"
          },
          "where": {
            "type": "object",
            "example": {
              "category_slug": "dla-seniorow"
            },
            "description": "Chroma DB metadata filtering expression"
          },
          "api_key": {
            "type": "string",
            "example": "YOUR_API_KEY",
            "description": "API Authorization Key"
          }
        }
      },
...
    }
  }
}